requirements.txt
# install libncurses5-dev if necessary
yowsup2